Introduce removeNonEmptyDirectory attribute for file systems (#1591)

+++ core/src/main/java/nl/nn/adapterframework/filesystem/LocalFileSystem.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2019, 2021 WeAreFrank!
+   Copyright 2019-2021 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
 */
 package nl.nn.adapterframework.filesystem;
 
+import java.io.File;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.DirectoryStream;
@@ -23,6 +24,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.nio.file.attribute.BasicFileAttributes;
+import java.util.Comparator;
 import java.util.Date;
 import java.util.Map;
 
@@ -141,10 +143,17 @@ public class LocalFileSystem extends FileSystemBase<Path> implements IWritableFi
 	}
 
 	@Override
-	public void removeFolder(String folder) throws FileSystemException {
+	public void removeFolder(String folder, boolean removeNonEmptyFolder) throws FileSystemException {
 		if (folderExists(folder)) {
 			try {
-				Files.delete(toFile(folder));
+				if(removeNonEmptyFolder) {
+					Files.walk(toFile(folder))
+						.sorted(Comparator.reverseOrder())
+						.map(Path::toFile)
+						.forEach(File::delete);
+				} else {
+					Files.delete(toFile(folder));
+				}
 			} catch (IOException e) {
 				throw new FileSystemException("Cannot remove folder ["+ folder +"]",e);
 			}
